--- a/python/sglang/srt/layers/quantization/fp8_utils.py
+++ b/python/sglang/srt/layers/quantization/fp8_utils.py
@@ -11,7 +11,7 @@ from sglang.srt.layers.quantization.fp8_kernel import (
 from sglang.srt.utils import get_bool_env_var, is_hip
 
 is_hip_ = is_hip()
-if is_hip_ and get_bool_env_var("AITER_BLOCK_GEMM"):
+if is_hip_ and get_bool_env_var("CK_MOE"):
     from aiter import gemm_a8w8_blockscale
 
 _is_cuda = torch.cuda.is_available() and torch.version.cuda
@@ -81,7 +81,7 @@ def apply_w8a8_block_fp8_linear(
         output = fp8_blockwise_scaled_mm(
             q_input, weight.T, x_scale, weight_scale.T, out_dtype=input.dtype
         )
-    elif is_hip_ and get_bool_env_var("AITER_BLOCK_GEMM"):
+    elif is_hip_ and get_bool_env_var("CK_MOE"):
         q_input, x_scale = per_token_group_quant_fp8(
             input_2d, block_size[1], column_major_scales=False
         )