<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation
Decode Performance boost (DSR1 +8%):
Before:
```
/sgl-workspace/sglang# RCCL_MSCCL_ENABLE=0 CK_MOE=1 python -m sglang.bench_one_batch --batch-size 64 --input 512 --output 32 --model deepseek-ai/DeepSeek-R1/ --tp 8 --trust-remote-code --quantization fp8
Benchmark ...
Prefill. latency: 2.02619 s, throughput:  16172.25 token/s
Decode.  latency: 0.03899 s, throughput:   1641.31 token/s
Decode.  latency: 0.03917 s, throughput:   1633.71 token/s
Decode.  latency: 0.03974 s, throughput:   1610.47 token/s
Decode.  latency: 0.04054 s, throughput:   1578.81 token/s
Decode.  latency: 0.04060 s, throughput:   1576.45 token/s
Decode.  median latency: 0.04177 s, median throughput:   1532.14 token/s
Total. latency:  3.311 s, throughput:  10514.95 token/s
```
After:
```
/sgl-workspace/sglang# RCCL_MSCCL_ENABLE=0 CK_MOE=1 AITER_BLOCK_GEMM=1 python -m sglang.bench_one_batch --batch-size 64 --input 512 --output 32 --model deepseek-ai/DeepSeek-R1/ --tp 8 --trust-remote-code --quantization fp8
Benchmark ...
Prefill. latency: 2.35484 s, throughput:  13915.15 token/s
Decode.  latency: 0.03659 s, throughput:   1749.00 token/s
Decode.  latency: 0.03680 s, throughput:   1739.00 token/s
Decode.  latency: 0.03729 s, throughput:   1716.12 token/s
Decode.  latency: 0.03764 s, throughput:   1700.34 token/s
Decode.  latency: 0.03790 s, throughput:   1688.75 token/s
Decode.  median latency: 0.03876 s, median throughput:   1651.03 token/s
Total. latency:  3.549 s, throughput:   9810.84 token/s
```
<!-- Explain the purpose of this PR and the goals it aims to achieve. -->

## Modifications
AITER BLOCK GEMM (tuned for DSR1/V3 for tp 8):
`AITER_BLOCK_GEMM=1` in command line or ENV to activate aiter/ck block gemm.
<!-- Describe the changes made in this PR. -->

## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.