@BruceXcluding 2025-03-05
> Let's just use `CK_MOE` to trigger it, no need `AITER_BLOCK_GEMM` to avoid deep tree of choices and many ENVs.

OK, modified.